requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.0
aiohttp==3.7.4.post0
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==3.0.1
attrify==3.0
attrs==24.2.0
braceexpand==0.1.7
Brotli==1.1.0
button_paginator @ git+https://github.com/andrewthederp/Button_paginator@17b7c3030d59881434730c16827cf9588e2226d3
certifi==2024.7.4
cffi==1.17.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@8686b5d1bbc1d3c862292eb436ab630d6e9c9b53
discord-py-paginators==0.2.1
discord.py==2.4.0
frozenlist==1.4.1
idna==2.10
import_expression==1.1.5
jishaku==2.5.2
multidict==6.0.5
pycares==4.4.0
pycparser==2.22
python-dotenv==1.0.1
reactionmenu==3.1.7
requests==2.32.3
saucenao-api==2.4.0
six==1.16.0
tracemoepy==4.1
typing_extensions==4.12.2
urllib3==2.2.2
wheel==0.44.0
yarl==1.9.4
